+++ scripts/build.sh
@@ -14,22 +14,21 @@ echo "Generating translations..."
 scripts/generate-mo.sh
 
 echo "Packing extension..."
-gnome-extensions pack \
+gnome-extensions pack src \
     --force \
     --extra-source="bin" \
     --extra-source="lib" \
     --extra-source="ui" \
-    --extra-source="LICENSE" \
-    --extra-source="README.md" \
-    --extra-source="CHANGELOG.md"
+    --extra-source="../LICENSE" \
+    --extra-source="../CHANGELOG.md"
 
 echo "Packing Done!"
 
 while getopts i flag; do
     case $flag in
 
-        i)  gnome-extensions install \
-            --force just-perfection-desktop@just-perfection.shell-extension.zip && \
+        i)  gnome-extensions install --force \
+            just-perfection-desktop@just-perfection.shell-extension.zip && \
             echo "Extension is installed. Now restart the GNOME Shell." || \
             { echo "ERROR: Could not install the extension!"; exit 1; };;