+++ scripts/generate-mo.sh
@@ -13,8 +13,8 @@ cd "$( cd "$( dirname "$0" )" && pwd )/.."
 for filename in po/*.po;
 do
     lang=$(basename "$filename" .po)
-    moPath="locale/$lang/LC_MESSAGES/just-perfection.mo"
-    mkdir -p "locale/$lang/LC_MESSAGES"
+    moPath="src/locale/$lang/LC_MESSAGES/just-perfection.mo"
+    mkdir -p "src/locale/$lang/LC_MESSAGES"
     msgfmt "$filename" --output-file="$moPath" && echo "$lang [OK]" || 
     	echo "ERROR: Failed to generate '$lang.po'."
 done