+++ CHANGELOG.md
@@ -18,6 +18,7 @@ We go to the next version after each release on [GNOME Shell Extensions website]
 - First swipe up in desktop startup status.
 - Window picker caption border on disable.
 - Galician translation by [@frandieguez](https://gitlab.gnome.org/frandieguez).
+- Spanish translation by [@DiegoIvanME](https://gitlab.gnome.org/DiegoIvanME).
 
 ### Added